+++ packages/ds/angular/src/common/jl-common.module.ts
@@ -1,0 +1,17 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { IconComponent } from './icon/icon.component';
+
+@NgModule({
+    declarations: [
+        IconComponent
+    ],
+    imports: [
+        CommonModule
+    ],
+    exports: [
+        IconComponent
+    ],
+    providers: [],
+})
+export class JlCommonModule { }