+++ packages/ds/angular/CHANGELOG.md
@@ -3,6 +3,62 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.2-beta.6](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.5...@ircc-ca/ds-sdc-angular@0.0.2-beta.6) (2022-04-12)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.5](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.4...@ircc-ca/ds-sdc-angular@0.0.2-beta.5) (2022-04-11)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.4](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.3...@ircc-ca/ds-sdc-angular@0.0.2-beta.4) (2022-04-05)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.3](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.2...@ircc-ca/ds-sdc-angular@0.0.2-beta.3) (2022-04-04)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.2](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.1...@ircc-ca/ds-sdc-angular@0.0.2-beta.2) (2022-04-04)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.1](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.2-beta.0...@ircc-ca/ds-sdc-angular@0.0.2-beta.1) (2022-04-01)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
+## [0.0.2-beta.0](https://github.com/IRCC-ca/ds-sdc/compare/@ircc-ca/ds-sdc-angular@0.0.1...@ircc-ca/ds-sdc-angular@0.0.2-beta.0) (2022-03-31)
+
+**Note:** Version bump only for package @ircc-ca/ds-sdc-angular
+
+
+
+
+
 ## 0.0.1 (2022-03-25)
 
 **Note:** Version bump only for package @ircc-ca/ds-sdc-angular